docker_requirements.txt
wheel
numpy==1.19.5
python3-opencv
torch==1.9.0
icecream==2.1.0
pytorch_lightning==1.3.0
tqdm==4.61.0
torchvision==0.10.0
Pillow==8.3.0
flask
pytorch_msssim